5.5.3 - Bug fixes to get_datafiles module.
BUG01: When the p2e parameter is used, the package root directory is
included in the destination path. However, the data (db scripts,
config files, etc.) should live in the root, along with the *.exe file.
FIX01: Updated the root path to exclude the package's root directory
name.
BUG02: README and LICENSE files were not being included.
FIX02: Improved functionality for gathering the README and LICENSE
files, and added parameters to get_datafiles() to either collect,
or ignore these files.